feat: manually edit mappings for BigQueryDataTransferConfig
